features: Fix ociMaxVersion
We are vendoring a runtime-spec version that breaks semver. Let's just
hardcode the ociMaxVersion, instead of using the version from the
bindings that reports "1.0.2-dev", that should be interpreted as
something bigger than "1.0.2".

That is not true according to semver, so let's just report "1.0.3-dev"
that is indeed bigger according to semver. We don't expect to change the
runtime-spec vendored version in the 1.1 release.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>