dmz: don't use runc-dmz in complicated capability setups
Due to the fact that runc-dmz is an intermediate binary without any
special set-capability file attributes, using runc-dmz for containers
with a non-root user can result in different capability sets being
applied after the second execve().

Linux capabilities are quite complicated, and there are loads of
different interactions between file and process capability sets, so we
should just go with the most conservative rule to determine if we can't
use runc-dmz -- if the inheritable, permitted, and bounding sets are not
equal to the ambient set then we don't use runc-dmz.

Fixes: dac417174654 ("runc-dmz: reduce memfd binary cloning cost with small C binary")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>