libct/cg/fs2: skip cpu.weight if cpu.idle is set
In case cpu.idle=1, cpu.weight can't be set (write to cpu.weights fails
with EINVAL). The systemd v2 cgroup driver already handles this case,
but the fs2 driver did not, resulting in an error when both cpu.idle
and cpu.weight is set.

Fix by not setting cpu.weight (with a warning).

Reported-by: axb <uaxb@hotmail.com>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>