Makefile: fixup for Go 1.23
It is no longer possible to compile runc with rlimit hack from
commit da68c8e3 with Go 1.23 (devel at the moment), it errors out:

> link: github.com/opencontainers/runc/libcontainer/system: invalid reference to syscall.origRlimitNofile

Go 1.23 implements some restrictions wrt accessing internal symbols,
and to work around those, we need to add a checklinkname=0 linker flag.

Alas, this flag is not supported in older versions, so the logic is
added to check go version and add the flag conditionally.

This fixes the issue.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>