let runc work with glibc lt 2.32 in go 1.22.x
In glibc versions older than 2.32 (before commit 4721f95058),
pthread_getattr_np does not always initialize the `attr` argument,
and when it fails, it results in a NULL pointer dereference in
pthread_attr_destroy down the road. This has been fixed in go 1.22.4.
We hack this to let runc can work with glibc < 2.32 in go 1.22.x,
once runc doesn't support 1.22, we can remove this hack.

Signed-off-by: lifubang <lifubang@acmcoder.com>